+++ tests/Feature/CanHandleFakePaymentTest.php
@@ -18,17 +18,17 @@ class CanHandleFakePaymentTest extends AbstractTestCase
         Event::fake();
         $this->paymentGateway = new FakePaymentGateway;
         $this->app->instance(PaymentGateway::class, $this->paymentGateway);
+
+        $this->order = new TestModel(['amount' => 12345]);
+        $this->order->save();
     }
 
     /** @test */
     public function canCreatePaymentRequest()
     {
-        $order = new TestModel(['amount' => 12345]);
-        $order->save();
-
         $payment = $this->paymentGateway->chargeAmountForPayable(
-            $order->amount,
-            $order,
+            $this->order->amount,
+            $this->order,
             'Some description',
             [ 'return_url' => 'http://fake-return-url.com' ]
         );
@@ -48,10 +48,7 @@ class CanHandleFakePaymentTest extends AbstractTestCase
     /** @test */
     public function canFetchPaymentUpdate()
     {
-        $order = new TestModel(['amount' => 12345]);
-        $order->save();
-
-        $payment = $order->payments()->save(new Payment([
+        $payment = $this->order->payments()->save(new Payment([
             'amount' => 12345,
             'status' => 'open',
             'description' => 'Some fake description',
@@ -64,7 +61,42 @@ class CanHandleFakePaymentTest extends AbstractTestCase
         $payment = $this->paymentGateway->fetchUpdateFor($payment);
         $this->assertEquals('paid', $payment->status);
         $this->assertNull($payment->redirect_url);
+    }
+
+    /** @test */
+    public function dispatchesPaymentUpdatedEventIfChanged()
+    {
+        $payment = $this->order->payments()->save(new Payment([
+            'amount' => 12345,
+            'status' => 'open',
+            'description' => 'Some fake description',
+            'return_url' => 'http://www.return-to-here.com',
+            'redirect_url' => 'http://www.redirect-to-here.com',
+            'gateway_name' => 'FakePaymentGateway',
+            'gateway_payment_reference' => 'fake-external-id-1',
+        ]));
+        
+        $payment = $this->paymentGateway->fetchUpdateFor($payment);
         Event::assertDispatched(PaymentUpdated::class, function ($e) use ($payment) {
+            return $e->payment->id === $payment->id;
+        });
+    }
+
+    /** @test */
+    public function doesNotdispatchPaymentUpdatedEventIfNotChanged()
+    {
+        $payment = $this->order->payments()->save(new Payment([
+            'amount' => 12345,
+            'status' => 'paid',
+            'description' => 'Some fake description',
+            'return_url' => 'http://www.return-to-here.com',
+            'redirect_url' => null,
+            'gateway_name' => 'FakePaymentGateway',
+            'gateway_payment_reference' => 'fake-external-id-1',
+        ]));
+        
+        $payment = $this->paymentGateway->fetchUpdateFor($payment);
+        Event::assertNotDispatched(PaymentUpdated::class, function ($e) use ($payment) {
             return $e->payment->id === $payment->id;
         });
     }